Added filter on item name in quotes
Filter added to item in quotes. 

Initially was thinking that is it needed to translate the quotes only? I can't find any languages that do this specifically, but if it's a RTL language, perhaps you'd swap around the quote direction.

The function `wc_add_to_cart_message` gives you the filter `wc_add_to_cart_message` to play with, but is already set in stone as HTML, and if you truly want to change the HTML of this message, you would need to rewrite the function.

Simply adding `<strong>` to the product title is not possible without preg_replacing or hooking into gettext