Attempt to reduce noise in PRs generated by codecov
This commit is an attempt to reduce the noise in PRs generated by the code coverage comment posted automatically by Codecov. It changes Codecov configuration to post a comment to a PR only if it changes the code coverage and it also removes reach graph, the diff image and the flags from the comment content. Leaving only a list of the files impacted by the PR. My initial plan was to leave in the comment content only a link to the full report, but I couldn't find a way to do that in the documentation (https://docs.codecov.io/docs/pull-request-comments).