Use entire sentence for checkout address_2 placeholder string
The placeholder string that is displayed in the address_2 field in the checkout was changed in #20639 to hide the "(optional)" part when the field is mandatory. The problem is that this was done splitting the sentence in two parts which goes against i18n best practices (https://developer.wordpress.org/themes/functionality/internationalization/#best-practices-for-writing-strings) as the order of the words can be different depending on the language. This commit creates two full sentences one with the "(optional)" part and another one without it to better support the translation of both strings.