'Used for variations' checkbox should only appear for variable products
When editing product attributes, the checkbox 'Used for variations' should only be displayed when editing a variable product. The code has checks in place to control the display of this checkbox when adding an attribute or when loading the product attributes, but it was missing a check when saving product attributes which is added in this commit.