Bump default WC API version used when creating webhook
This commit bumps the default WC API version used when users are creating a new webhook. Before this change the default version would be v2, and after this change it is v3 which is the latest API version.

In the future we might want to create a method in WC_API class or somewhere else that returns the latest WC API version and use it in `includes/class-wc-webhook.php` instead of hard-coding the version number.