Update npm packages to their latest versions
This commit updates npm packages to their latest versions:

- chromedriver  2.46.0  →  74.0.0
- husky          1.3.1  →   2.1.0
- node-sass     4.11.0  →  4.12.0

Those three packages are devDependencies and since we are early in the WC 3.7 release cycle this change shouldn't cause any issue. I'm doing this to see if updating the chromedriver package fixes the failing e2e test that is failing on Travis (example of a failure: https://travis-ci.org/woocommerce/woocommerce/jobs/524016366#L829). I'm not able to reproduce this error on my local machine and in the past updating chromedriver fixed failing e2e tests.