Delete images that are not used anymore
This commit deletes the files assets/images/klarna-white.png and
assets/images/square-white.png. In #23690 we stopped using those two
files so there is no reason to keep them in the repository. Since #23690
was included in a patch release, we opted to keep the images for a
little longer and remove them in a minor release.