Remove code duplication from OBW payment gateway logic
In #23690 we implemented a few changes to the logic to decide which
payment gateways to display in the onboarding wizard. When working on
this code, we noticed that there was a bit of duplication in it that was
not necessary (see
https://github.com/woocommerce/woocommerce/pull/23690#discussion_r284004924).
We opted to remove this duplication in a different commit, since that PR
was included in a patch release.