Remove shipping section from helpful message of login form
Shipping section may not always be present such as in case of only
virtual products in cart, so remove it from message shown for the
login form.
Fixes #23940 by removing shipping section from message. In case of
where shipping section needs to be shown, it is quite obvious for
customers to move to shipping section after filling out billing details.