Remove deprecated i18n/states directory
This commit removes the deprecated i18n/states directory. Since PR #22339,
more specifically commit f1afbcf03a704f86fed0e5e7ef153ea0be9d900c,
the list of states for each country is defined in i18n/states.php and the
directory i18n/states/ is not used anymore. It is not clear from the PR
and commit descriptions if i18n/states/ was left on purpose or not.
I'm assuming it was not on purpose as it doesn't seem to be used
anywhere.